backend/requirements.txt
PyMuPDF>=1.24.0
openai>=1.12.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.9
